+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 2025-05-03
+
+### Added
+
+- `no_date` flag for turning off the date string in the entry
+
+### Changed
+
+- Renamed `hide_level_string` to `no_level_string`
+
 ## [0.1.4] - 2025-05-03
 
 ### Added